nikame/scaffold/templates/auth/jwt/requirements.txt
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.0.0
